Create Request Policie

+++ src/main/java/com/company/insurance_request/domain/model/Coverage.java
@@ -3,10 +3,14 @@ package com.company.insurance_request.domain.model;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class Coverage {
-    private Double roubo;
-    private Double perdaTotal;
-    private Double colisaoComTerceiros;
+
+    //todo - Padronizar objeto
+    private BigDecimal roubo;
+    private BigDecimal perdaTotal;
+    private BigDecimal colisaoComTerceiros;
 }
